requirements.txt
beautifulsoup4==4.7.1
dialogflow==0.5.2
Flask==1.0.2
gunicorn==19.9.0
lxml==4.3.0
requests==2.21.0